refactor: Move Config and fetch_votes to query.py
