+++ tools/cfp-vote-counter/main.py
@@ -1,60 +1,13 @@
 #!/usr/bin/env python3
-from datetime import datetime
-from decimal import Decimal
-from typing import Literal, Final
+from typing import Literal
 
 import fire
 import pandas as pd
 import matplotlib.pyplot as plt
 
-import boto3
-from boto3.dynamodb.conditions import Key
-
+from src.query import QueryConfig, fetch_votes
 from src.transform import unique_over_time, count_votes, time_series_total_count
-from src.types import DynamoResponse, Col
 
-
-DYNAMO_VOTE_TABLE_PRD: Final[str] = "voteCFP-prd-VoteTableC0BC27A7-UKB7XFRIUIX1"
-DYNAMO_VOTE_TABLE_STG: Final[str] = "voteCFP-stg-VoteTableC0BC27A7-84BXPDSTU937"
-
-VOTING_PERIOD: Final[dict[str, tuple[datetime, datetime]]] = {
-    "cndt2022": (datetime.fromisoformat('2022-10-01'), datetime.fromisoformat('2022-10-13T18:00:00Z+09:00'))
-}
-
-
-class Config:
-    cfp_votetable: str
-    event_name: str
-    voting_start: Decimal
-    voting_end: Decimal
-
-    def __init__(self, event: str, env: str):
-        if env == "prd":
-            self.cfp_votetable = DYNAMO_VOTE_TABLE_PRD
-        elif env == "stg":
-            self.cfp_votetable = DYNAMO_VOTE_TABLE_STG
-        else:
-            raise ValueError(f"env not exist: name={env}")
-
-        if event not in VOTING_PERIOD:
-            raise ValueError(f"event not exist: name={event}")
-        self.event_name = event
-        self.voting_start = Decimal(VOTING_PERIOD[event][0].timestamp() * 1000)
-        self.voting_end = Decimal(VOTING_PERIOD[event][1].timestamp() * 1000)
-
-
-def fetch_votes(conf: Config) -> DynamoResponse:
-    dynamo = boto3.resource("dynamodb")
-    res: DynamoResponse = dynamo.Table(conf.cfp_votetable).query(
-        KeyConditionExpression=(
-                Key(Col.EVENT_NAME).eq(conf.event_name)
-                & Key(Col.TIMESTAMP).between(conf.voting_start, conf.voting_end)
-        ),
-        ProjectionExpression=f"#{Col.TIMESTAMP}, {Col.GLOBAL_IP}, {Col.TALK_ID}",
-        ExpressionAttributeNames={f"#{Col.TIMESTAMP}": Col.TIMESTAMP},
-        Limit=10000,
-    )
-    return res
 
 class Command:
     """CFP Vote Counter CLI"""
@@ -67,7 +20,7 @@ class Command:
         :param env: Environment (stg|prd)
         :param span: Seconds of span where multiple votes from the same GIP would be considered as the same one
         """
-        conf = Config(event, env)
+        conf = QueryConfig(event, env)
         res = fetch_votes(conf)
         if res["Count"] == 0:
             print("No votes found.")
@@ -86,7 +39,7 @@ class Command:
         :param env: Environment (stg|prd)
         :param span: Seconds of span where multiple votes from the same GIP would be considered as the same one
         """
-        conf = Config(event, env)
+        conf = QueryConfig(event, env)
         res = fetch_votes(conf)
         if res["Count"] == 0:
             print("No votes found.")
